backend/requirements.txt
fastapi[all]>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
Pillow>=10.0.0
apscheduler>=3.10.0
aiosmtplib>=3.0.0
email-validator>=2.0.0
